ajuste da arquitetura
