+++ src/frontend/.vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "dd60942e",
+  "configHash": "b4ee949e",
+  "lockfileHash": "a5548aea",
+  "browserHash": "6c2ba803",
+  "optimized": {},
+  "chunks": {}
+}
