+++ src/frontend/src/app/components/FSMStatus.tsx
